fix db usage
