+++ packages/core/utils/package.json
@@ -32,6 +32,7 @@
     "expect-type": "^0.20.0",
     "express": "^4.21.0",
     "jest": "^29.7.0",
+    "medusa-test-utils": "^1.1.44",
     "rimraf": "^5.0.1",
     "typescript": "^5.6.2"
   },
